+++ lib/slack-notifier.rb
@@ -7,20 +7,27 @@ require_relative 'slack-notifier/link_formatter'
 
 module Slack
   class Notifier
-    attr_reader :endpoint, :http_client, :default_payload
+    attr_reader :endpoint, :default_payload
 
     def initialize webhook_url, options={}
       @endpoint        = URI.parse webhook_url
-      @http_client     = options.delete(:http_client) || DefaultHTTPClient
       @default_payload = options
     end
 
     def ping message, options={}
-      message = LinkFormatter.format(message)
-      payload = { text: message }.merge(default_payload).merge(options)
+      message      = LinkFormatter.format(message)
+      payload      = default_payload.merge(options).merge(text: message)
+      client       = payload.delete(:http_client) || http_client
+      http_options = payload.delete(:http_options)
 
+      params = { payload: payload.to_json }
+      params[:http_options] = http_options if http_options
 
-      http_client.post endpoint, payload: payload.to_json
+      client.post endpoint, params
+    end
+
+    def http_client
+      default_payload.fetch :http_client, DefaultHTTPClient
     end
 
     def channel